requirements.txt
requests
beautifulsoup4
pyautogui
pillow
openai
python-dotenv
selenium
pyperclip
pywinauto
pytesseract
charset-normalizer
pydantic
opencv-python
numpy
tensorflow
scipy
fuzzywuzzy
pygame
anthropic
winocr
asyncio

# To be cleaned up